fix unfair coin position reset
It was possible for the red agent to be on the coin after the coin being moved (if two following time the coin position match with the blue, success is not reset to 0 and the sum reach 2). This provided the red agent with a small advantage, enough to prevent any cooperation.